backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
qdrant-client==1.7.0
openai==1.3.5
pydantic==2.5.0
python-multipart==0.0.6
sentence-transformers==2.2.2